Add food text and images. Extract data
