+++ src/app/industry-use-cases/page.tsx
@@ -16,20 +16,8 @@ import Plus from "@/svgs/plus.svg";
 import Minus from "@/svgs/minus.svg";
 import { Button } from "@/components/ui/button";
 import { AnimatePresence, motion } from "framer-motion";
-import AccordionItemContent from "@/components/industry-use-cases/accordion-item-content";
-import type { AccordionItemsContentType } from "@/components/industry-use-cases/accordion-item-content";
-
-type TriggerImagesAndText = {
-  text1: string;
-  imageSrc1: string;
-  text2: string;
-  imageSrc2: string;
-};
-
-type AccordionContent = {
-  id: string;
-  content?: ReactNode;
-} & TriggerImagesAndText;
+import { ACCORDION_ITEMS } from "./data";
+import { AccordionContentType, TriggerImagesAndText } from "./types";
 
 const MotionPlus = motion(Plus);
 const MotionMinus = motion(Minus);
@@ -39,7 +27,7 @@ const Item = ({
   content,
   active,
 }: {
-  triggerContent: AccordionContent;
+  triggerContent: AccordionContentType;
   content?: ReactNode;
   active: boolean;
 }) => {
@@ -130,142 +118,6 @@ const TriggerContent = ({
   );
 };
 
-const ACCORDION_ITEMS_CONTENT: AccordionItemsContentType = {
-  constructions: {
-    ecosystem: {
-      content1: (
-        <>
-          <div className="text-xl tracking-tight">
-            Why are forests essential for businesses and society?
-          </div>
-          <p className="max-w-[685px] tracking-tight">
-            Forests are indispensable to society&apos;s efforts to tackle climate change. They act
-            as the Earth&apos;s lungs, absorbing enormous amounts of carbon dioxide and replenishing
-            the oxygen in the atmosphere. Forests also provide vital raw materials like timber,
-            paper, and biofuels, supporting key construction, manufacturing, and energy industries.
-          </p>
-          <p className="max-w-[685px] tracking-tight">
-            Conservation of the world&apos;s biodiversity is inextricably tied to forests. Forests
-            cover only 31% of the global land area but support 80% of the world&apos;s amphibian
-            species, 75% of bird species, and 68% of mammals.
-          </p>
-        </>
-      ),
-      image1: "/assets/industry-cases-constructions-ecosystem-1.png",
-      imageText:
-        "Conservation of the world's biodiversity is inextricably tied to forests. Forests cover only 31% of the global land area but support 80% of the world's amphibian species, 75% of bird species, and 68% of mammals.",
-      content2: (
-        <>
-          <div className="mb-5 text-xl">
-            Forests provide goods and services to businesses and society.
-          </div>
-          <div className="flex flex-col gap-5">
-            <p>
-              Forests supply essential raw materials such as timber and minerals, as well as genetic
-              material for pharmaceuticals, food, and fibre. They also support crucial ecosystem
-              services like climate regulation and water filtration and offer recreational
-              opportunities that boost tourism.
-            </p>
-            <p>
-              Forests are culturally significant, enriching communities with places for reflection,
-              relaxation, and spiritual renewal. They also support traditional practices and social
-              spaces.
-            </p>
-            <p>
-              These benefits underscore the importance of sustainable forest management for
-              economic, cultural and environmental well-being.
-            </p>
-          </div>
-        </>
-      ),
-      image2: "/assets/industry-cases-constructions-ecosystem-2.png",
-      directIndustries: [
-        {
-          title: "Manufacturing and construction",
-          tooltip: "For raw materials (timber, minerals, clay) used in manufacturing of products.",
-        },
-        {
-          title: "Building and furniture design",
-          tooltip:
-            "Forests for timber, a fundamental material for creating structural components and furniture.",
-        },
-        {
-          title: "Mining",
-          tooltip:
-            "Significant reserves of many minerals (e.g., bauxite, cobalt, gold, nickel, copper, tin, iron ore) are found beneath forest ecosystems.",
-        },
-        {
-          title: "Tourism, recreation and hospitality",
-          tooltip:
-            "For scenic beauty and natural landscapes, which attract visitors and enhance outdoor experiences.",
-        },
-      ],
-      indirectIndustries: [
-        {
-          title: "Food and beverage manufacturers",
-          tooltip:
-            "For raw materials like paper and packaging, and ingredients such as fruits, nuts, herbs and spices.",
-        },
-        {
-          title: "Fibre and clothing manufacturers",
-          tooltip:
-            "For wood and plant-based fibres, essential for producing sustainable textiles and packaging materials.",
-        },
-        {
-          title: "Retail",
-          tooltip:
-            "For packaging materials and paper products, which are essential for product presentation and distribution.",
-        },
-        {
-          title: "Technology",
-          tooltip:
-            "For rare minerals like lithium and cobalt for batteries and electronics, and biomaterials like cellulose for paper and biodegradable plastics.",
-        },
-      ],
-      image3: "/assets/industry-cases-constructions-ecosystem-3.png",
-      insights: [
-        "Essential Raw Materials: Forests provide timber and minerals for construction, making sustainable forest management crucial for reliable material supply.",
-        "Vital Ecosystem Services: Forests support construction through climate regulation, carbon storage, and soil and water quality, which are critical for sustainable operations.",
-        "Deforestation Risks: Deforestation can cause biodiversity loss and degrade ecosystem services, leading to operational, financial, and reputational risks for the construction industry.",
-      ],
-    },
-    dependencies: {},
-    impacts: {},
-  },
-};
-const ACCORDION_ITEMS: AccordionContent[] = [
-  {
-    id: "Constructions",
-    text1: "Manufacturing and construction",
-    text2: "depend on Forests.",
-    imageSrc1: "/assets/industry-cases-accordion-1.png",
-    imageSrc2: "/assets/industry-cases-accordion-2.png",
-    content: <AccordionItemContent {...ACCORDION_ITEMS_CONTENT["constructions"]} />,
-  },
-  {
-    id: "Tourism",
-    text1: "Tourism and Hospitality",
-    text2: "depend on Coastal Ecosystems.",
-    imageSrc1: "/assets/industry-cases-accordion-3.png",
-    imageSrc2: "/assets/industry-cases-accordion-4.png",
-    content: <div>Byyeeee</div>,
-  },
-  {
-    id: "Urban Planning",
-    text1: "Retail depends on",
-    text2: "Urban Green Spaces.",
-    imageSrc1: "/assets/industry-cases-accordion-5.png",
-    imageSrc2: "/assets/industry-cases-accordion-6.png",
-  },
-  {
-    id: "Food",
-    text1: "Food Security depends on",
-    text2: "Ecosystem Assets.",
-    imageSrc1: "/assets/industry-cases-accordion-7.png",
-    imageSrc2: "/assets/industry-cases-accordion-8.png",
-  },
-];
-
 const IndustryUseCasesPage: FC = () => {
   const [activeItem, setActiveItem] = useState<string>("");
 
@@ -290,7 +142,7 @@ const IndustryUseCasesPage: FC = () => {
               "lg:opacity-100": !!activeItem,
             })}
           >
-            {ACCORDION_ITEMS.map((accordionItem: AccordionContent) => (
+            {ACCORDION_ITEMS.map((accordionItem: AccordionContentType) => (
               <li key={accordionItem.id}>
                 <Button
                   type="button"
@@ -318,7 +170,7 @@ const IndustryUseCasesPage: FC = () => {
             className="flex flex-col lg:gap-y-20"
             collapsible
           >
-            {ACCORDION_ITEMS.map((accordionItem: AccordionContent) => (
+            {ACCORDION_ITEMS.map((accordionItem: AccordionContentType) => (
               <Item
                 key={accordionItem.id}
                 triggerContent={accordionItem}
